+++ src/layouts/SharedLayout.jsx
@@ -1,0 +1,17 @@
+import { Outlet } from 'react-router-dom';
+import Header from '../components/Header/Header';
+import Footer from '../components/Footer/Footer';
+
+const SharedLayout = () => {
+  return (
+    <>
+      <Header />
+      <main>
+        <Outlet />
+      </main>
+      <Footer />
+    </>
+  );
+};
+
+export default SharedLayout;
