+++ src/pages/BodyCarePage/BodyCarePage.jsx
