+++ src/pages/HomePage/HomePage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Intro } from '../../components/Intro/Intro';
 import { Achievemets } from '../../components/Achievements/Achievemets';
 
-import { MainRouteStyled } from './MainRouteStyled';
+import { MainRouteStyled } from './HomePageStyled';
 import Services from '../../components/Services/Services';
 import AboutMain from '../../components/AboutMain/AboutMain';
 import PriceMain from '../../components/PriceMain/PriceMain';
