+++ src/pages/PricePage/PricePage.jsx
