+++ src/routes/routesPath.js
@@ -1,0 +1,8 @@
+export const routesPath = {
+  HOME_PAGE: ``,
+  BODY_CARE: `body_care`,
+  ARTTHERAPY: `arttherapy`,
+  ABOUT: `about`,
+  EVENTS: `events`,
+  PRICE: `price`,
+};
