+++ vite.config.js
@@ -41,6 +41,7 @@ export default defineConfig({
       "@services": path.resolve(__dirname, "src/services"),
       "@store": path.resolve(__dirname, "src/store"),
       "@utils": path.resolve(__dirname, "src/utils"),
+      "@routes": path.resolve(__dirname, "src/routes"),
     },
   },
 });
